refactor: 자율회 게시글 작성 뷰 이동
